+++ ricaun.RevitTest.Command/Command/IRunTestService.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace ricaun.RevitTest.Command
+{
+    public interface IRunTestService
+    {
+        public string[] GetTests(string filePath);
+        public bool RunTests(
+            string fileToTest,
+            int revitVersionNumber,
+            Action<string> actionOutput = null,
+            bool forceToOpenNewRevit = false,
+            bool forceToWaitRevit = false,
+            bool forceToCloseRevit = false,
+            params string[] testFilters);
+    }
+}