+++ ricaun.RevitTest.Console/Revit/RevitRunTestService.cs
@@ -1,4 +1,5 @@
 ﻿using ricaun.RevitTest.Console.Command;
+using ricaun.RevitTest.Console.Revit.Utils;
 using System;
 
 namespace ricaun.RevitTest.Console.Revit